Change method SumSales and fix variables

+++ src/test/java/ru/netology/StatsService/services/StatsServiceTest.java
@@ -29,8 +29,8 @@ public class StatsServiceTest {
     public void shouldCalculateSum() {
         StatsService service = new StatsService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 180;
-        int actualSum = service.sumSales(sales);
+        long expectedSum = 180;
+        long actualSum = service.sumSales(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
     }
@@ -39,8 +39,8 @@ public class StatsServiceTest {
     public void shouldCalculateAverage() {
         StatsService service = new StatsService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedAverage = 15;
-        int actualAverage = service.averageSales(sales);
+        long expectedAverage = 15;
+        long actualAverage = service.averageSales(sales);
 
         Assertions.assertEquals(expectedAverage, actualAverage);
     }
